% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx-error.R
\name{approx_error}
\alias{approx_error}
\title{Compute Approximation Errors}
\usage{
approx_error(operator, f, x)
}
\arguments{
\item{operator}{An object of class \code{"approx_operator"}.}

\item{f}{Function to be approximated.}

\item{x}{Numeric vector of evaluation points.}
}
\value{
An object of class \code{"approx_error"} and
  \code{"data.frame"} containing the evaluation points, exact
  function values, approximated values, signed errors, absolute
  errors, and relative errors.
}
\description{
Computes pointwise approximation errors for an approximation
operator at one or more evaluation points.
}
\details{
For an operator L_n and a function f, the signed error is

  L_n(f; x) - f(x),

while the absolute and relative errors are

  |L_n(f; x) - f(x)|

and

  |L_n(f; x) - f(x)| / |f(x)|,

respectively. If f(x) = 0, the relative error is returned as NA.
}
