% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{approximate}
\alias{approximate}
\title{Evaluate an Approximation Operator}
\usage{
approximate(operator, f, x)
}
\arguments{
\item{operator}{An object of class \code{"approx_operator"}.}

\item{f}{Function to be approximated.}

\item{x}{Numeric vector of evaluation points.}
}
\value{
A numeric vector containing the approximated values.
}
\description{
Evaluates a positive linear approximation operator for a given function
at one or more points.
}
