% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-table.R
\name{error_table}
\alias{error_table}
\title{Create an Approximation Error Table}
\usage{
error_table(operator, f, x, n = NULL, digits = 6)
}
\arguments{
\item{operator}{An object of class \code{"approx_operator"}.}

\item{f}{Function to be approximated.}

\item{x}{Numeric vector of evaluation points.}

\item{n}{Optional vector of positive integers. If \code{NULL},
the value of \code{n} stored in \code{operator} is used.}

\item{digits}{Number of digits used when printing the table.}
}
\value{
An object of class \code{"approx_error_table"} and
  \code{"data.frame"} containing pointwise absolute approximation
  errors.
}
\description{
Creates a numerical error table for an approximation operator.
The table may be computed for the operator's current value of n
or for several values of n.
}
