% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-approx-parameter-scan.R
\name{plot.approx_parameter_scan}
\alias{plot.approx_parameter_scan}
\title{Plot an Approximation Parameter Scan}
\usage{
\method{plot}{approx_parameter_scan}(
  x,
  metric = c("max_absolute_error", "mae", "rmse"),
  type = "b",
  lwd = NULL,
  pch = NULL,
  show_values = TRUE,
  grid_lines = TRUE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"approx_parameter_scan"}.}

\item{metric}{Error metric to be plotted. One of
\code{"max_absolute_error"}, \code{"mae"}, or \code{"rmse"}.}

\item{type}{Plot type used to display the parameter scan.}

\item{lwd}{Optional line width. If \code{NULL}, the ApproxOp
graphical default is used.}

\item{pch}{Optional plotting symbol. If \code{NULL}, the ApproxOp
graphical default is used.}

\item{show_values}{Logical. If \code{TRUE}, the scanned parameter
values are displayed explicitly on the horizontal axis.}

\item{grid_lines}{Logical. If \code{TRUE}, light grid lines are
displayed.}

\item{xlab}{Optional label for the horizontal axis. If \code{NULL},
the scanned parameter name is used.}

\item{ylab}{Optional label for the vertical axis.}

\item{main}{Optional plot title.}

\item{...}{Additional graphical arguments passed to
\code{graphics::plot()}.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Produces a graphical representation of approximation error as a
function of a scanned operator parameter.
}
\details{
The available error metrics are the maximum absolute error,
mean absolute error (MAE), and root mean squared error (RMSE).
}
\examples{
C <- approx_operator(
  family = "sheffer",
  subfamily = "charlier",
  n = 20,
  params = list(a = 2)
)

Scan <- parameter_scan(
  C,
  parameter = "a",
  values = c(1.5, 2, 3, 5, 10),
  f = function(x) x^2,
  grid = seq(0, 1, length.out = 101)
)

plot(Scan)
plot(Scan, metric = "mae")

}
