% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-error-table.R
\name{print.approx_error_table}
\alias{print.approx_error_table}
\title{Print an Approximation Error Table}
\usage{
\method{print}{approx_error_table}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"approx_error_table"}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Prints an object of class \code{"approx_error_table"} in a
structured table with horizontal and vertical separators.
}
