% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-approx-operator.R
\name{print.approx_operator}
\alias{print.approx_operator}
\title{Print an Approximation Operator}
\usage{
\method{print}{approx_operator}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"approx_operator"}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The operator object, invisibly.
}
\description{
Prints a concise summary of an approximation operator object.
}
