% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-approx-operator-check.R
\name{print.approx_operator_check}
\alias{print.approx_operator_check}
\title{Print an Approximation Operator Check}
\usage{
\method{print}{approx_operator_check}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"approx_operator_check"}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Prints a compact summary of the diagnostic results produced by
\code{check_operator()}.
}
